requirements.txt
Django==2.0.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
isort==4.3.4
Markdown==2.6.11
pylint==1.8.2
pylint-django==0.9.1
pylint-plugin-utils==0.2.6
raven==6.9.0
redgreenunittest==0.1.1
Scrapy==1.5.0
scrapydo==0.2.2
whitenoise==3.3.1
